Change fields types for money and notes types of fields

+++ manage_app/models.py
@@ -5,7 +5,7 @@ from company_panel.models import Company
 class Task(models.Model):
     company = models.ForeignKey(Company, models.CASCADE, related_name='+')
     name = models.CharField(max_length=150)
-    default_hourly_rate = models.FloatField(default=0.0)
+    default_hourly_rate = models.DecimalField(default=0, max_digits=20, decimal_places=2)
 
     def __str__(self):
         return self.name
